refactor: separate funnel as client page component
